requirements.txt
grpcio-tools
pydub
sounddevice
soundfile
PyAudio
vosk
numpy

fuzzywuzzy
python-Levenshtein

